trace-cmd: Do not print stacks after stacks
When a filter is in place, the stack trace for the events that are
printed by the filter should also be printed. Stack traces for events
that are filtered out, should also be filtered out. But there was a bug
in the logic that checked if the last event was printed or not to know
to print the stack, and that is, the stack from the printed event was
considered a printed event itself. So a stack trace coming from another
event could be considered "printed" if two stacks came back to back
(which can happen because of interrupts.

Note, this does mean that logic should be added to test for interrupts.
But that's another story.

Link: https://lore.kernel.org/linux-trace-devel/20240528113141.7a3e8a04@rorschach.local.home

Reported-by: Vlastimil Babka <vbabka@suse.cz>
Fixes: 82ed4a937 ("trace-cmd library: Add filtering logic for iterating events")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
Tested-by: Vlastimil Babka <vbabka@suse.cz>