libtracecmd: Optimize what cpus to check in tracecmd_iterate_events()
Currently, tracecmd_iterate_events() will look at all possible CPUs to
find the next event to print, even if the CPUs to look at are filtered.
Instead, create a list of all the CPUs to look at and iterate that, such
that only the CPUs that are needed are looked at instead of all of them,
including those that are filtered out.

Link: https://lore.kernel.org/linux-trace-devel/20240528145219.7cef53b0@rorschach.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>