trace-cmd lib: Prevent possible memory corruption in add_plugin_file()
If strdup() fails the error path access original address of
pdata->files after it has been dereferenced. Make sure that
pdata->files contains the up-to-date pointer before calling calling a
function that could fail.

This was flagged as resource leak (CWE-772) because ptr isn't freed
in that scenario, but there is something worse going on that the
static analysis missed.

Link: https://lore.kernel.org/linux-trace-devel/20240605134054.2626953-6-jmarchan@redhat.com

Signed-off-by: Jerome Marchand <jmarchan@redhat.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>