trace-cmd record: Always disable func_stack_trace in tracecmd_disable_all_tracing()
The function tracecmd_disable_all_tracing() is passed a boolean parameter
to decide to disable the tracers or not. If the tracers are not to be
disabled (as it resets the trace file) it also does not disable the
func_stack_trace. Unfortunately, that means the set_ftrace_filter gets
reset when the function stack tracing is still enabled, causing the end of
the trace to take a long time.

Always disable func_stack_trace option in tracecmd_disable_all_tracing().

Link: https://lore.kernel.org/20240809212152.6486337d@rorschach.local.home
Fixes: 935d5c5a033a5 ("trace-cmd reset: Remove functions in function tracing filters")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>