trace-cmd record: Fix stdin redirection to /dev/null
In daemonize_start(), stdin file descriptor is closed after the call
to dup2. This doesn't make sense and could lead to FD zero being
reused. I assume the original intend was to close the devnull FD which
in the current code remains opened while the devnull variable goes out
of scope.

Close devnull instead of 0.

Fixes: a3929b09b4f65 ("trace-cmd record: Add --daemonize")
Link: https://lore.kernel.org/20241205144439.127564-8-jmarchan@redhat.com
Signed-off-by: Jerome Marchand <jmarchan@redhat.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>