trace-cmd: Add sqlhist to bash completion
Add the commands for sqlhist to bash completion. This makes writing the SQL
lines for the select statements so much easier. Now the user can use tab
completion to see what the next possible commands are.

This also makes the shortcuts like TIMESTAMP_DELTA_USECS more visible and
less error prone to typos.

Link: https://lore.kernel.org/20250409224350.7224d8ce@gandalf.local.home
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>