trace-cmd list: Show all functions that match string by default
Before tracefs_filter_functions() was used to find the list of functions,
the trace-cmd list -f would list all functions that match the given string
by default. After the switching over to tracefs_filter_functions(), by
default it does a exact match.

This is quite annoying, as in most cases the user wants to see what
available functions have the given string. If the user knows the function
they are looking for, they do not need to use trace-cmd list -f !

Link: https://lore.kernel.org/20250411093628.3f68fb02@gandalf.local.home
Fixes: 39acb4cc1 ("trace-cmd list: Use tracefs_filter_functions()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>