trace-cmd completion: Fix trace-cmd report to show files
The addition of the trace-cmd sqlhist had to separate out where commands
are printed and where files are printed, and it split cmd_options into one
that showed the commands and cmd_options_files that also will show the
files if no command was found to be completed.

Some trace-cmd commands should show files and some should not.

trace-cmd report is one that should show files if nothing else is
completed, but it was given cmd_options when it should have used
cmd_options_files.

Link: https://lore.kernel.org/20250416201249.195f3e86@gandalf.local.home
Fixes: 2149da9cdc48b ("trace-cmd: Add sqlhist to bash completion")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>