trace-cmd split: Only open one file descriptor per CPU
When splitting a trace.dat file, trace-cmd split will create a temp file
per CPU as it parse the trace.dat to store what to save in the new
trace.dat file into per CPU files.

After it finishes the creation of the CPU data files, it then calls
tracecmd_append_cpu_data() which will open each of these CPU files and
append them to the new trace.dat file.

The issue is when you have a trace.dat file that was created by a machine
with 512 CPUs. Because it doesn't close the file descriptors after
creating the files and before calling tracecmd_append_cpu_data() it has
1024 file descriptors open. With some distributions having a file
descriptor limit of 1024, this exceeds the limit because there's other
file descriptors open (for the trace.dat files themselves), and the
program fails.

There's no reason to keep the file descriptors open for the files that
were created before calling tracecmd_append_cpu_data(). Just close them
before calling that function. This keeps the number of file descriptors
open to that of the number of CPU buffers in the trace.dat file instead of
doubling them. There is a couple more file descriptors open as mentioned
before, but those are not an issue.

Cc: Julia Lawall <Julia.Lawall@inria.fr>
Link: https://lore.kernel.org/20250708143819.09124bd7@gandalf.local.home
Reported-by: Julia Lawall <julia.lawall@inria.fr>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>