libtracecmd: Fix memory leak on partial reverse iteration
When calling tracecmd_iterate_events_reverse with a callback that does
not always return 0, the trace is only partially iterated. By that, the
non-iterated records are leaked, resulting in the error:

 1 pages still allocated on cpu <cpu>

We fix this by always iterating the remaining events on all selected
CPUs. In the full iteration case, this stops on the first record as this
is already zero. In the partial iteration case, all remaining records
are freed, which is - by construction of the records list - at max a
page size.

Link: https://lore.kernel.org/20251121134749.1530855-1-felix.moessbauer@siemens.com
Fixes: 56cbc522518a5 ("trace-cmd library: Add tracecmd_iterate_events_reverse() API")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>