libtracecmd: Fix continuing in tracecmd_iterate_events_reverse()
When a callback from tracecmd_iterate_events_reverse() returns non-zero it
cause the function to exit out early. Calling tracecmd_iterate_events_reverse()
again with cont=true, is supposed to restart where it left off. But
because of the way the reverse traverses all the events on a page, by
calling tracecmd_read_data() to find the next task, it does not pick the
next task. It can pick the next task at the start of the page.

If the callback back causes the function to exit out early, record the
page_offset of the last record read, and then set the cursor back to that
event. Then the next tracecmd_iterate_events_reverse() called with
continue will start with the next record after the one that was last
read.

Link: https://lore.kernel.org/all/20251121120117.20e82d9e@gandalf.local.home/

Cc: Felix Moessbauer <felix.moessbauer@siemens.com>
Link: https://lore.kernel.org/20251203220353.4c520524@robin
Fixes: 56cbc522518a5 ("trace-cmd library: Add tracecmd_iterate_events_reverse() API")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>